Refactoring Testing Traits
