+++ tests/TestCase.php
@@ -2,9 +2,13 @@
 
 namespace Tests;
 
+use Tests\Traits\CreatesApplication;
 use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
 
+/**
+ * 
+ */
 abstract class TestCase extends BaseTestCase
 {
-    use CreatesApplication, CreatesUsers;
+    use CreatesApplication;
 }
